Add .test.ts extension to tests. Add timeouts.
