--- packages/create-next-stack/src/tests/e2e/tests/css-modules-with-sass/css-modules-with-sass-only.ts
@@ -1,11 +1,0 @@
-import { test } from "@jest/globals"
-import { testArgsWithFinalChecks } from "../../helpers/test-args"
-
-test("testCssModulesWithSassOnly", async () => {
-  testArgsWithFinalChecks([
-    "--debug",
-    "--package-manager=pnpm",
-    "--styling=css-modules-with-sass",
-    ".",
-  ])
-})
