+++ packages/create-next-stack/src/tests/e2e/tests/css-modules/css-modules-only.test.ts
@@ -1,0 +1,15 @@
+import { test } from "@jest/globals"
+import { testArgsWithFinalChecks } from "../../helpers/test-args"
+
+test(
+  "testCssModulesOnly",
+  async () => {
+    testArgsWithFinalChecks([
+      "--debug",
+      "--package-manager=pnpm",
+      "--styling=css-modules",
+      ".",
+    ])
+  },
+  10 * 60 * 1000
+)
