--- packages/create-next-stack/src/tests/e2e/tests/css-modules/css-modules-only.ts
@@ -1,11 +1,0 @@
-import { test } from "@jest/globals"
-import { testArgsWithFinalChecks } from "../../helpers/test-args"
-
-test("testCssModulesOnly", async () => {
-  testArgsWithFinalChecks([
-    "--debug",
-    "--package-manager=pnpm",
-    "--styling=css-modules",
-    ".",
-  ])
-})
