+++ packages/create-next-stack/src/tests/e2e/tests/emotion/emotion-all-flags.test.ts
@@ -1,0 +1,25 @@
+import { test } from "@jest/globals"
+import { testArgsWithFinalChecks } from "../../helpers/test-args"
+
+test(
+  "testEmotionAllFlags",
+  async () => {
+    testArgsWithFinalChecks([
+      "--debug",
+      "--package-manager=pnpm",
+      "--styling=emotion",
+      "--mantine",
+      "--chakra",
+      "--material-ui",
+      "--react-hook-form",
+      "--formik",
+      "--framer-motion",
+      "--prettier",
+      "--formatting-pre-commit-hook",
+      "--react-icons",
+      "--react-query",
+      ".",
+    ])
+  },
+  10 * 60 * 1000
+)
