--- packages/create-next-stack/src/tests/e2e/tests/emotion/emotion-all-flags.ts
@@ -1,21 +1,0 @@
-import { test } from "@jest/globals"
-import { testArgsWithFinalChecks } from "../../helpers/test-args"
-
-test("testEmotionAllFlags", async () => {
-  testArgsWithFinalChecks([
-    "--debug",
-    "--package-manager=pnpm",
-    "--styling=emotion",
-    "--mantine",
-    "--chakra",
-    "--material-ui",
-    "--react-hook-form",
-    "--formik",
-    "--framer-motion",
-    "--prettier",
-    "--formatting-pre-commit-hook",
-    "--react-icons",
-    "--react-query",
-    ".",
-  ])
-})
