+++ packages/create-next-stack/src/tests/e2e/tests/emotion/emotion-only.test.ts
@@ -1,0 +1,15 @@
+import { test } from "@jest/globals"
+import { testArgsWithFinalChecks } from "../../helpers/test-args"
+
+test(
+  "testEmotionOnly",
+  async () => {
+    testArgsWithFinalChecks([
+      "--debug",
+      "--package-manager=pnpm",
+      "--styling=emotion",
+      ".",
+    ])
+  },
+  10 * 60 * 1000
+)
