--- packages/create-next-stack/src/tests/e2e/tests/emotion/emotion-only.ts
@@ -1,11 +1,0 @@
-import { test } from "@jest/globals"
-import { testArgsWithFinalChecks } from "../../helpers/test-args"
-
-test("testEmotionOnly", async () => {
-  testArgsWithFinalChecks([
-    "--debug",
-    "--package-manager=pnpm",
-    "--styling=emotion",
-    ".",
-  ])
-})
