+++ packages/create-next-stack/src/tests/e2e/tests/styled-components/styled-components-all-flags.test.ts
@@ -1,0 +1,22 @@
+import { test } from "@jest/globals"
+import { testArgsWithFinalChecks } from "../../helpers/test-args"
+
+test(
+  "testStyledComponentsAllFlags",
+  async () => {
+    testArgsWithFinalChecks([
+      "--debug",
+      "--package-manager=pnpm",
+      "--prettier",
+      "--styling=styled-components",
+      "--react-hook-form",
+      "--formik",
+      "--framer-motion",
+      "--formatting-pre-commit-hook",
+      "--react-icons",
+      "--react-query",
+      ".",
+    ])
+  },
+  10 * 60 * 1000
+)
