+++ packages/create-next-stack/src/tests/e2e/tests/styled-components/styled-components-only.test.ts
@@ -1,0 +1,15 @@
+import { test } from "@jest/globals"
+import { testArgsWithFinalChecks } from "../../helpers/test-args"
+
+test(
+  "testStyledComponentsOnly",
+  async () => {
+    testArgsWithFinalChecks([
+      "--debug",
+      "--package-manager=pnpm",
+      "--styling=styled-components",
+      ".",
+    ])
+  },
+  10 * 60 * 1000
+)
