--- packages/create-next-stack/src/tests/e2e/tests/styled-components/styled-components-only.ts
@@ -1,11 +1,0 @@
-import { test } from "@jest/globals"
-import { testArgsWithFinalChecks } from "../../helpers/test-args"
-
-test("testStyledComponentsOnly", async () => {
-  testArgsWithFinalChecks([
-    "--debug",
-    "--package-manager=pnpm",
-    "--styling=styled-components",
-    ".",
-  ])
-})
