+++ packages/create-next-stack/src/tests/e2e/tests/tailwind-css/tailwind-css-only.test.ts
@@ -1,0 +1,15 @@
+import { test } from "@jest/globals"
+import { testArgsWithFinalChecks } from "../../helpers/test-args"
+
+test(
+  "testTailwindCssOnly",
+  async () => {
+    testArgsWithFinalChecks([
+      "--debug",
+      "--package-manager=pnpm",
+      "--styling=tailwind-css",
+      ".",
+    ])
+  },
+  10 * 60 * 1000
+)
