--- packages/create-next-stack/src/tests/e2e/tests/tailwind-css/tailwind-css-only.ts
@@ -1,11 +1,0 @@
-import { test } from "@jest/globals"
-import { testArgsWithFinalChecks } from "../../helpers/test-args"
-
-test("testTailwindCssOnly", async () => {
-  testArgsWithFinalChecks([
-    "--debug",
-    "--package-manager=pnpm",
-    "--styling=tailwind-css",
-    ".",
-  ])
-})
