+++ packages/create-next-stack/src/tests/e2e/tests/test-help-and-version-flags.test.ts
@@ -1,0 +1,12 @@
+import { describe, test } from "@jest/globals"
+import { testArgsWithoutFinalChecks } from "../helpers/test-args"
+
+describe("testHelpAndVersionFlags", () => {
+  test("testHelpFlag", async () => {
+    testArgsWithoutFinalChecks(["--help"])
+  })
+
+  test("testVersionFlag", async () => {
+    testArgsWithoutFinalChecks(["--version"])
+  })
+})
