--- packages/create-next-stack/src/tests/e2e/tests/test-help-and-version-flags.ts
@@ -1,10 +1,0 @@
-import { test } from "@jest/globals"
-import { testArgsWithoutFinalChecks } from "../helpers/test-args"
-
-test("testHelpFlag", async () => {
-  testArgsWithoutFinalChecks(["--help"])
-})
-
-test("testVersionFlag", async () => {
-  testArgsWithoutFinalChecks(["--version"])
-})
