--- packages/create-next-stack/src/tests/e2e/tests/test-yarn.ts
@@ -1,31 +1,0 @@
-import { expect, test } from "@jest/globals"
-import { exists } from "../../../main/helpers/exists"
-import { testArgsWithFinalChecks } from "../helpers/test-args"
-
-test("testYarn", async () => {
-  const { runDirectory } = await testArgsWithFinalChecks([
-    "--debug",
-    "--package-manager=yarn",
-    "--styling=emotion",
-    "--mantine",
-    "--chakra",
-    "--material-ui",
-    "--react-hook-form",
-    "--formik",
-    "--framer-motion",
-    "--prettier",
-    "--formatting-pre-commit-hook",
-    "--react-icons",
-    "--react-query",
-    ".",
-  ])
-
-  const yarnLockExists = await exists(`${runDirectory}/yarn.lock`)
-  expect(yarnLockExists).toBe(true)
-
-  const packageLockExists = await exists(`${runDirectory}/package-lock.json`)
-  expect(packageLockExists).toBe(false)
-
-  const pnpmLockExists = await exists(`${runDirectory}/pnpm-lock.yaml`)
-  expect(pnpmLockExists).toBe(false)
-})
